pyproject.toml
[tool.poetry]
name = "personalassistant-ai-bot"
version = "0.1.0"
description = ""
authors = ["d43v30n <anjey.mail@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
openai = "^1.3.5"
ics = "^0.7.2"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
aiosqlite = "^0.19.0"
isort = "^5.12.0"
pytesseract = "^0.3.10"
opencv-python-headless = "^4.8.1.78"
uvicorn = "^0.25.0"
fastapi = "^0.108.0"
jinja2 = "^3.1.2"
asyncpg = "^0.29.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
